refactor(mcp): dedupe receipt handlers, extract SQL helpers, section TranscriptIndex
Codebase audit 2026-07-08 wave 2, spec W2-E. Pure structural motion in
Tools/TranscriptedMCP, no behavior change and no tool-surface change
(verified byte-identical tools/list output before/after):

- ToolHandlers.swift (1676 lines) split by tool family into
  ToolHandlers+Meetings/+Dictations/+Search/+Rollups/+Receipts.swift, with
  the tool schema registration + call dispatch left in ToolHandlers.swift.
- handleDecisions/handleCommitments/handleOpenQuestions/handleSearchMeetings
  were near-identical copy-paste on their result-building tail (build
  CrossMeetingToolResult, fire telemetry if non-empty, encode). Factored
  that tail into a private handleReceiptQuery helper in
  ToolHandlers+Receipts.swift; each tool keeps its own thin wrapper doing
  its own index query + item-to-receipt mapping (which do differ), so the
  four MCP tool entry points are unchanged.
- TranscriptIndex.swift's SQL plumbing (SQLBinding enum, bind/bindExec/
  exec/execOrThrow, colText/colTextOptional/dbError) extracted into
  SQLiteHelpers.swift as free functions taking an explicit `db` parameter,
  since TranscriptIndex and EmbeddingStore each own an independent sqlite3
  connection. TranscriptIndex's private methods now delegate to the shared
  functions so every existing call site is unchanged. EmbeddingStore's
  colText/dbErrorLocked (byte-identical logic) now delegate too;
  execLocked and the raw prepare/bind loops were left as-is where they mix
  in loop/transaction control flow or intentionally distinct log text, to
  avoid a behavior change beyond mechanical dedup.
- TranscriptIndex's ~190-line createTables() DDL block moved to
  TranscriptIndex+Schema.swift as an extension; `exec` bumped from private
  to internal so the cross-file extension can call it (private is
  file-scoped in Swift).

Verification: swift build && swift test in Tools/TranscriptedMCP (171
tests, 0 failures); captured tools/list output before the refactor and
diffed it against tools/list after — all 18 tool names and full
schemas/descriptions are byte-identical; smoke-tested status/decisions/
commitments/open_questions/search_meetings over stdio, all return
well-formed results with isError: false.

Deviations from spec: EmbeddingStore's raw sqlite3_prepare_v2/bind loops
(~lines 117-338) were only partially deduped (colText/dbErrorLocked) — the
rest mixes read-then-write batching, blob binding, and silent
error-swallowing that isn't a 1:1 mechanical match for the shared throwing/
logging helpers, so converting them risked a real behavior change
(added logs, different error propagation) rather than pure motion.
TranscriptIndex+Schema.swift covers the DDL (createTables); the schema-
version-gate/migration bookkeeping (applySchemaVersionGate, storedUserVersion)
was left in TranscriptIndex.swift since it's tightly coupled to `db`/
`indexPath` lifecycle, not just DDL text.
