+++ Tools/TranscriptedMCP/Sources/TranscriptedMCP/ToolHandlers+Meetings.swift
@@ -1,0 +1,234 @@
+import Foundation
+import MCP
+import TranscriptedCaptureKit
+
+// MARK: - list_meetings
+
+func handleListMeetings(params: CallTool.Parameters, index: TranscriptIndex, meetingDirs: [URL]) throws -> CallTool.Result {
+    let count = params.arguments?["count"]?.intValue ?? 10
+    let date = params.arguments?["date"]?.stringValue
+    let dateFrom = params.arguments?["date_from"]?.stringValue ?? date
+    let dateTo = params.arguments?["date_to"]?.stringValue ?? date
+
+    var results = try index.listMeetings(count: count, dateFrom: dateFrom, dateTo: dateTo)
+
+    // Populate titles from markdown YAML frontmatter
+    for i in results.indices {
+        guard case .valid(let mdURL) = PathSecurity.resolveReadableFile(
+            named: results[i].filename,
+            appendingExtension: "md",
+            in: meetingDirs
+        ) else { continue }
+        if let content = CaptureMarkdown.readBoundedContents(of: mdURL) {
+            results[i].title = extractTitle(from: content) ?? results[i].filename
+        }
+    }
+
+    if results.isEmpty {
+        return try emptyResult(scope: .meetings, searchedDirectories: meetingDirs, index: index)
+    }
+
+    trackAgentCaptureQueryObserved(
+        toolKind: "list",
+        captureKind: "meeting",
+        sourceCount: results.count
+    )
+
+    let json = try JSONEncoder.pretty.encode(results)
+    return textResult(String(data: json, encoding: .utf8) ?? "[]")
+}
+
+/// Extract title from markdown YAML frontmatter
+func extractTitle(from content: String) -> String? {
+    CaptureMarkdown.extractTitle(from: content)
+}
+
+/// Returns non-empty transcript dialogue lines from the `## Full Transcript` section, filtering boilerplate.
+func extractDialogueLines(from content: String) -> [String] {
+    let markers = ["## Full Transcript\n", "## Transcript\n"]
+    guard let range = markers.compactMap({ content.range(of: $0) }).first else { return [] }
+    return String(content[range.upperBound...])
+        .components(separatedBy: "\n")
+        .filter { !$0.isEmpty && !$0.hasPrefix("*Generated by") && $0 != "---" }
+}
+
+// MARK: - read_meeting
+
+func handleReadMeeting(params: CallTool.Parameters, meetingDirs: [URL]) throws -> CallTool.Result {
+    guard let filename = params.arguments?["filename"]?.stringValue, !filename.isEmpty else {
+        return textResult("Missing required parameter: filename", isError: true)
+    }
+
+    let section = params.arguments?["section"]?.stringValue ?? "full"
+    let offset = max(0, params.arguments?["offset"]?.intValue ?? 0)
+    let limit = params.arguments?["limit"]?.intValue
+    let paginationRequested = limit != nil || offset > 0
+
+    let mdURL: URL
+    switch PathSecurity.resolveReadableFile(named: filename, appendingExtension: "md", in: meetingDirs) {
+    case .valid(let url):
+        mdURL = url
+    case .missing:
+        return textResult("Meeting not found: \(filename). Use list_meetings to see available meetings.", isError: true)
+    case .invalid:
+        return textResult("Invalid filename: \(filename)", isError: true)
+    }
+
+    guard let content = CaptureMarkdown.readBoundedContents(of: mdURL) else {
+        return textResult("Meeting not found: \(filename). Use list_meetings to see available meetings.", isError: true)
+    }
+
+    let parsed = CaptureMarkdownParser.parseMeeting(from: content)
+
+    trackAgentCaptureQueryObserved(
+        toolKind: "read",
+        captureKind: "meeting",
+        sourceCount: 1
+    )
+
+    switch section {
+    case "transcript":
+        let dialogue = extractDialogueLines(from: content).joined(separator: "\n").trimmingCharacters(in: .whitespacesAndNewlines)
+        let raw = dialogue.isEmpty ? content : dialogue
+        if !paginationRequested, raw.count <= maxUnpaginatedReadCharacters {
+            return textResult(raw)
+        }
+        return try meetingTranscriptPageResult(
+            parsed: parsed, content: content, filename: filename,
+            offset: offset, limit: limit, includeFrontmatter: false, rawFallback: raw
+        )
+
+    case "speakers":
+        // Return YAML frontmatter + speaker analytics section
+        var result = ""
+        if let frontmatter = frontmatterBlock(of: content) {
+            result += frontmatter
+        }
+        if let analyticsRange = content.range(of: "## Channel & Speaker Analytics") {
+            if let transcriptRange = content.range(of: "## Full Transcript") {
+                result += "\n" + String(content[analyticsRange.lowerBound..<transcriptRange.lowerBound])
+            } else {
+                result += "\n" + String(content[analyticsRange.lowerBound...])
+            }
+        }
+        return textResult(result.isEmpty ? content : result)
+
+    default: // "full"
+        if !paginationRequested, content.count <= maxUnpaginatedReadCharacters {
+            return textResult(content)
+        }
+        return try meetingTranscriptPageResult(
+            parsed: parsed, content: content, filename: filename,
+            offset: offset, limit: limit, includeFrontmatter: true, rawFallback: content
+        )
+    }
+}
+
+/// Bounded transcript read: frontmatter metadata plus one utterance window and
+/// explicit pagination fields, so a long meeting never comes back as an
+/// unbounded dump. Falls back to the raw markdown when the file does not parse
+/// as a windowable transcript — we can only page what the parser understands.
+private func meetingTranscriptPageResult(
+    parsed: ParsedMeetingCapture?,
+    content: String,
+    filename: String,
+    offset: Int,
+    limit: Int?,
+    includeFrontmatter: Bool,
+    rawFallback: String
+) throws -> CallTool.Result {
+    guard let parsed, !parsed.utterances.isEmpty else {
+        return textResult(rawFallback)
+    }
+
+    let total = parsed.utterances.count
+    let start = min(offset, total)
+    let end: Int
+    if let limit {
+        end = min(start + max(1, limit), total)
+    } else {
+        end = autoWindowEnd(items: parsed.utterances, start: start) {
+            $0.text.count + paginationItemOverheadCharacters
+        }
+    }
+
+    let speakerNames = Dictionary(uniqueKeysWithValues: parsed.speakers.map { ($0.id, $0.name) })
+    let utterances = parsed.utterances[start..<end].map { utterance in
+        MeetingTranscriptPageUtterance(
+            start: utterance.start,
+            end: utterance.end,
+            speaker: speakerNames[utterance.speakerId] ?? utterance.speakerId,
+            speakerId: utterance.speakerId,
+            text: utterance.text
+        )
+    }
+
+    let returned = utterances.count
+    let truncated = start + returned < total
+    let nextOffset = truncated ? start + returned : nil
+
+    let hint: String
+    if offset >= total {
+        hint = "offset \(offset) is past the end — this transcript has \(total) utterances. Retry with a smaller offset."
+    } else if let nextOffset {
+        hint = "Showing utterances \(start)-\(end - 1) of \(total). Call read_meeting again with offset=\(nextOffset) to continue, or use section \"speakers\" for the overview without dialogue."
+    } else {
+        hint = "End of transcript — all remaining utterances included."
+    }
+
+    let page = MeetingTranscriptPage(
+        filename: filename,
+        frontmatter: includeFrontmatter ? frontmatterBlock(of: content) : nil,
+        totalUtterances: total,
+        offset: offset,
+        returned: returned,
+        truncated: truncated,
+        nextOffset: nextOffset,
+        hint: hint,
+        utterances: utterances
+    )
+
+    let json = try JSONEncoder.pretty.encode(page)
+    return textResult(String(data: json, encoding: .utf8) ?? "{}")
+}
+
+/// Raw YAML frontmatter block, matching the exact slice the speakers section
+/// has always returned (both fences plus the character after the closing one).
+private func frontmatterBlock(of content: String) -> String? {
+    guard content.count >= 8, content.hasPrefix("---"),
+          let endRange = content.range(of: "\n---\n", range: content.index(content.startIndex, offsetBy: 3)..<content.endIndex) else {
+        return nil
+    }
+    return String(content[content.startIndex...endRange.upperBound])
+}
+
+// MARK: - Meeting title hydration (shared across Search / Rollups / Receipts)
+
+/// Frontmatter title for a meeting markdown file, or nil when the file is
+/// unreadable or its frontmatter has no title.
+func frontmatterMeetingTitle(for filename: String, meetingDirs: [URL]) -> String? {
+    guard case .valid(let mdURL) = PathSecurity.resolveReadableFile(
+        named: filename,
+        appendingExtension: "md",
+        in: meetingDirs
+    ),
+    let content = CaptureMarkdown.readBoundedContents(of: mdURL) else {
+        return nil
+    }
+    return extractTitle(from: content)
+}
+
+func meetingTitle(for filename: String, meetingDirs: [URL]) -> String {
+    frontmatterMeetingTitle(for: filename, meetingDirs: meetingDirs) ?? filename
+}
+
+/// Replace the index's filename-derived meeting titles with real frontmatter
+/// titles. Groups whose markdown has no frontmatter title keep the
+/// filename-derived title they already carry.
+func hydrateMeetingSearchTitles(in results: inout GroupedSearchResult, meetingDirs: [URL]) {
+    for index in results.results.indices {
+        if let title = frontmatterMeetingTitle(for: results.results[index].filename, meetingDirs: meetingDirs) {
+            results.results[index].meetingTitle = title
+        }
+    }
+}
